The OrnsteinZernike Module

OrnsteinZernike — Module

A generic solver package for Ornstein-Zernike equations from liquid state theory

Module Index

OrnsteinZernike.DensityRamp

OrnsteinZernike.Exact

OrnsteinZernike.FourierIteration

OrnsteinZernike.HypernettedChain

OrnsteinZernike.MeanSphericalApproximation

OrnsteinZernike.MultiComponentHardSpheres

OrnsteinZernike.NgIteration

OrnsteinZernike.OZSolution

OrnsteinZernike.PercusYevick

OrnsteinZernike.SimpleLiquid

OrnsteinZernike.SingleComponentHardSpheres

OrnsteinZernike.compute_compressibility

OrnsteinZernike.compute_excess_energy

OrnsteinZernike.compute_virial_pressure

OrnsteinZernike.solve

Detailed API

OrnsteinZernike.DensityRamp — Type

DensityRamp <: Method

Solves the system by iteratively solving systems of increasing density, using the previous solution as initial guess at a higher density. This may help deal with convergence issues

Arguments

method: method by which to solve the system for individual densities.

densities: densities to consider. Must be a vector of increasing values.

verbose: whether to print information.

Example: DensityRamp(NgIteration(), [0.1, 0.3, 0.4]; verbose=false)

OrnsteinZernike.Exact — Type

Exact <: Method

Solves the system exactly. This is only implemented for specific systems.

Construct using Exact(; M=2^10, dr = sqrt(π/(M+1))/(2π)) Here, M is the number of points that the exact solution is evaluated on, and dr is the grid spacing. These are used to perform fourier transformations.

Examples

method = Exact() method = Exact(M=1000) method = Exact(M=1000, dr=0.01)

OrnsteinZernike.FourierIteration — Type

FourierIteration <: Method

Solves the system by recursive iteration in Fourier Space. Essentially, the algorithm is:

guess an initial γ(r)

find c(r) using the closure relation

fourier transform to get ĉ(k)

find γ(k) using the OZ-eq in k-space

compute γ(r) with a inverse fourier transform

compare with previous value, if not converged go to 2.

Optionally, a mixing rule is used to mix the new and previous iteration of c(r) in step 2.

Arguments:

M::Int: number of points discretize the solution on

dr::Float64: grid spacing in real space

mixing_parameter::Float64: mixing parameter for iteration mixing. A value of 1 is no mixing. Must be between 0 and 1.

max_iterations::Int64: maximal number of iterations

tolerance::Float64: tolerance to be reached

verbose::Bool: whether or not to print convergence information

default: FourierIteration(; mixingparameter=0.5, maxiterations=10^5, tolerance=10^-6, verbose=true, M=2^10, dr=sqrt(π/(M+1))/(2π))

OrnsteinZernike.HypernettedChain — Type

HypernettedChain

Implements the Hypernetted Chain closure c(r) = (f(r)+1)*exp(γ(r)) - γ(r) - 1, or equivalently b(r) = 0.

Example: closure = HypernettedChain()

OrnsteinZernike.MeanSphericalApproximation — Type

MeanSphericalApproximation

Implements the MSA closure c(r) = -βu(r), or equivalently b(r) = ln(γ(r) - βu(r) + 1) - γ(r) + βu(r).

Example: closure = MeanSphericalApproximation()

OrnsteinZernike.MultiComponentHardSpheres — Type

MultiComponentHardSpheres

Implements the hard-sphere pair interaction uᵢⱼ(r) = inf for r < Dᵢⱼ and uᵢⱼ(r) = 0 for r > Dᵢⱼ for a multicomponent system.

Expects a vector Dᵢ of diameters for each of the species. An additive mixing rule is used (Dᵢⱼ = (Dᵢ+Dⱼ)/2).

Example: closure = MultiComponentHardSpheres([0.8, 0.9, 1.0])

OrnsteinZernike.NgIteration — Type

NgIteration <: Method

Solves the system by recursive iteration in Fourier Space, and uses the Ng acceleration method. Essentially, the algorithm is:

guess an initial γ(r)

find c(r) using the closure relation

fourier transform to get ĉ(k)

find γ(k) using the OZ-eq in k-space

compute γ(r) with a inverse fourier transform

use Ng's method to provide a next guess for γ

compare with previous value, if not converged go to 2.

Arguments:

M::Int: number of points discretize the solution on

dr::Float64: grid spacing in real space

N_stages::Int: Number of previous values to take into account for step 6. A higher number should lead to faster convergence, yet more computation time per iteration.

max_iterations::Int64: maximal number of iterations

tolerance::Float64: tolerance to be reached

verbose::Bool: whether or not to print convergence information

default: NgIteration(; Nstages=3, maxiterations=10^3, tolerance=10^-6, verbose=true, M=2^10, dr=sqrt(π/(M+1))/(2π))

OrnsteinZernike.OZSolution — Type

OZSolution

Holds the solution of an Ornstein Zernike problem.

Fields:

r: vector of distances

k: vector of wave numbers

gr: radial distribution function

Sk: static structure factor

ck: direct correlation function in k space

cr: direct correlation function in real space

if the system was a single-component system, gr, Sk, ck and cr are vectors. If instead the system was a multicomponent one, they are three dimensional vectors, where the first dimension contains the values along r, and the second and third dimension contain the data for the species.

OrnsteinZernike.PercusYevick — Type

PercusYevick

Implements the Percus-Yevick closure c(r) = f(r)*(1+γ(r)), or equivalently b(r) = ln(1 + γ(r)) - γ(r).

Example: closure = PercusYevick()

OrnsteinZernike.SimpleLiquid — Type

SimpleLiquid{dims, ...} <: System

Holds information about a homogeneous, isotropic system with radially symmetric pair interactions. dims is the dimensionality.

Construct using

SimpleLiquid(dims, ρ, kBT, potential)

Fields:

ρ: number density, must be either a Number in case of a single component system, or a Vector in case of a mixture. In the latter case, each element contains the number density of the respective component.

kBT: thermal energy

potential::Potential: the interaction potential.

Examples:

ρ = 0.5; kBT = 1.1; dims = 3
pot = SingleComponentHardSpheres()
system = SimpleLiquid(dims, ρ, kBT, pot)

ρ = [0.5, 0.1]; kBT = 5.2; dims = 3
pot = MultiComponentHardSpheres([1.0, 0.8])
system = SimpleLiquid(dims, ρ, kBT, pot)

OrnsteinZernike.SingleComponentHardSpheres — Type

SingleComponentHardSpheres

Implements the hard-sphere pair interaction u(r) = inf for r < 1 and u(r) = 0 for r > 1.

Example: closure = SingleComponentHardSpheres()

OrnsteinZernike.compute_compressibility — Method

compute_compressibility(sol::OZSolution, system::SimpleLiquid)

Computes the isothermal compressibility χ of the system

uses the formula 1/χ = 1 - ρ ĉ(k=0) for single component systems and 1/χ = 1 - ρ Σᵢⱼ xᵢxⱼ ĉᵢⱼ(k=0) for mixtures.

OrnsteinZernike.compute_excess_energy — Method

compute_excess_energy(sol::OZSolution, system::SimpleLiquid)

Computes the excess energy per particle Eₓ, such that E = (dims/2kBT + Eₓ)N.

uses the formula Eₓ = 1/2 ρ ∫dr g(r) u(r) for single component systems and Eₓ = 1/2 ρ Σᵢⱼ xᵢxⱼ ∫dr gᵢⱼ(r) uᵢⱼ(r) for mixtures. Here x is the concentration fraction ρᵢ/sum(ρ).

OrnsteinZernike.compute_virial_pressure — Method

compute_virial_pressure(sol::OZSolution, system::SimpleLiquid)

Computes the pressure via the virial route

uses the formula p = kBTρ - 1/6 ρ^2 ∫dr g(r) u'(r) for single component systems and p = kBT Σᵢρᵢ - 1/6 Σᵢ ρᵢρⱼ ∫dr gᵢⱼ(r) u'ᵢⱼ(r) for mixtures.

OrnsteinZernike.solve — Function

solve(system::SimpleLiquid, closure::Closure, method::Method)

Solves the system system using the closure closure with method method.

solve(system::SimpleLiquid, closure::Closure)

Solves the system system using the closure closure with the default method NgIteration().