Closures

This package defines several closure relations that can be used out of the box. It is straightforward to implement your own closure, see Defining your own closure.

Implemented Closures

Below is a list of implemented closures. We use the notation shown in the Theory section.

OrnsteinZernike.HypernettedChain — Type

HypernettedChain

Implements the Hypernetted Chain closure $c(r) = (f(r)+1)\exp(\gamma(r)) - \gamma(r) - 1$, or equivalently $b(r) = 0$.

Example:

closure = HypernettedChain()

OrnsteinZernike.MeanSphericalApproximation — Type

MeanSphericalApproximation

Implements the MSA closure $c(r) = -\beta u(r)$, or equivalently $b(r) = \ln(\gamma(r) - \beta u(r) + 1) - γ(r) + \beta u(r)$.

Example:

closure = MeanSphericalApproximation()

OrnsteinZernike.PercusYevick — Type

PercusYevick

Implements the Percus-Yevick closure $c(r) = f(r)(1+\gamma(r))$, or equivalently $b(r) = \ln(1 + \gamma(r)) - γ(r)$.

Example:

closure = PercusYevick()