Interaction Potentials

This package defines several potentials that can be used out of the box. It is straightforward to implement your own potential, see Defining your own potentials. To evaluate the potential, call OrnsteinZernike.evaluate_potential(potential, r). For example:

using OrnsteinZernike, Plots
r = 0.9:0.01:4.0
potential = SingleComponentLennardJones(1.0, 1.0)
u = OrnsteinZernike.evaluate_potential(potential, r)
plot(r, u, xlabel="r", ylabel="u(r)", ylims=(-1,1), label=nothing)

Implemented interaction potentials

Below is a list of implemented closures. We use the notation shown in the Theory section.

OrnsteinZernike.MultiComponentHardSpheres — Type

MultiComponentHardSpheres

Implements the hard-sphere pair interaction $u_{ij}(r) = \infty$ for $r < D_{ij}$ and $u_{ij}(r) = 0$ for $r > D_{ij}$.

Expects a vector $D_i$ of diameters for each of the species. An additive mixing rule is used $\left(D_{ij} = (D_{i}+D_{j})/2\right)$.

Example:

closure = MultiComponentHardSpheres([0.8, 0.9, 1.0])

OrnsteinZernike.SingleComponentHardSpheres — Type

SingleComponentHardSpheres

Implements the hard-sphere pair interaction $ u(r) = \infty$ for $r < 1$ and $u(r) = 0$ for $r > 1$.

Example:

closure = SingleComponentHardSpheres()

OrnsteinZernike.SingleComponentLennardJones — Type

SingleComponentLennardJones

Implements the Lennard-Jones pair interaction $u(r) = 4\epsilon [(\sigma/r)^{12} - (\sigma/r)^6]$.

Expects values ϵ and σ, which respecively are the strength of the potential and particle size.

Example:

closure = SingleComponentLennardJones(1.0, 2.0)