Solvers

Having defined a SimpleLiquid and a Closure, one should choose a method by which to solve the equations. The implemented methods are Exact, FourierIteration, NgIteration.

All solvers in some way, need to define a grid on which to solve the equations. This is done using the keyword arguments for M and dr, which represent respectively the number of grid points, and the spacing between them. Some solvers have additional settings, such as a tolerance. Using powers of 2 for M typically give the best performance (for 3D systems).

For the implemented cases, Exact solves the system exactly, or throws an error if the method is not implemented.

OrnsteinZernike.Exact — Type

Exact <: Method

Solves the system exactly. This is only implemented for specific systems.

Right now, the implemented exact methods are

three-dimensional single-component system of hard spheres with the Percus Yevick closure

three-dimensional multi-component system of hard spheres with the Percus Yevick closure

Construct using

Exact(; M=2^10, dr = sqrt(π/(M+1))/(2π))

Here, M is the number of points that the exact solution is evaluated on, and dr is the grid spacing. These are used to perform Fourier transformations.

Examples

method = Exact()

method = Exact(M=1000)

method = Exact(M=1000, dr=0.01)

The methods FourierIteration and NgIteration both use recursive iteration to find improved estimates of the solution using Fourier Transforms. NgIteration uses a scheme to accelerate convergence, see Ref. [1].

OrnsteinZernike.FourierIteration — Type

FourierIteration <: Method

Solves the system by recursive iteration in Fourier Space. Essentially, the algorithm is:

guess an initial γ(r)

find c(r) using the closure relation

fourier transform to get ĉ(k)

find γ(k) using the OZ-eq in k-space

compute γ(r) with a inverse fourier transform

compare with previous value, if not converged go to 2.

Optionally, a mixing rule is used to mix the new and previous iteration of c(r) in step 2.

Arguments:

M::Int: number of points discretize the solution on

dr::Float64: grid spacing in real space

mixing_parameter::Float64: mixing parameter for iteration mixing. A value of 1 is no mixing. Must be between 0 and 1.

max_iterations::Int64: maximal number of iterations

tolerance::Float64: tolerance to be reached

verbose::Bool: whether or not to print convergence information

Default: FourierIteration(; mixing_parameter=0.5, max_iterations=10^5, tolerance=10^-6, verbose=true, M=2^10, dr=sqrt(π/(M+1))/(2π))

OrnsteinZernike.NgIteration — Type

NgIteration <: Method

Solves the system by recursive iteration in Fourier Space, and uses the Ng acceleration method. Essentially, the algorithm is:

guess an initial γ(r)

find c(r) using the closure relation

fourier transform to get ĉ(k)

find γ(k) using the OZ-eq in k-space

compute γ(r) with a inverse fourier transform

use Ng's method to provide a next guess for γ

compare with previous value, if not converged go to 2.

Arguments:

M::Int: number of points discretize the solution on

dr::Float64: grid spacing in real space

N_stages::Int: Number of previous values to take into account for step 6. A higher number should lead to faster convergence, yet more computation time per iteration.

max_iterations::Int64: maximal number of iterations

tolerance::Float64: tolerance to be reached

verbose::Bool: whether or not to print convergence information

Default: NgIteration(; N_stages=3, max_iterations=10^3, tolerance=10^-6, verbose=true, M=2^10, dr=sqrt(π/(M+1))/(2π))

References: Ng, K. C. (1974). Hypernetted chain solutions for the classical one‐component plasma up to Γ= 7000. The Journal of Chemical Physics, 61(7), 2680-2689.