The OrnsteinZernike Module

Missing docstring.

Missing docstring for OrnsteinZernike. Check Documenter's build log for details.

Module Index

OrnsteinZernike.find_mayer_f_function

OrnsteinZernike.fourier!

OrnsteinZernike.inverse_fourier!

OrnsteinZernike.solve

OrnsteinZernike.solve

Detailed API

OrnsteinZernike.find_mayer_f_function — Method

exp(- beta * u) - 1.

OrnsteinZernike.fourier! — Method

fourier!(F̂, F, plan, equation::OZEquation{3, T1, T2, T3}) where {T1, T2, T3}

computes the three dimensional radial fourier transform of F = rf(r), returning F̂ = kf̂(k), where f̂ is the fourier transform of f. it uses the discrete sine tranform provided by the r2r function of FFTW internally.

OrnsteinZernike.inverse_fourier! — Method

inverse_fourier!(F, F̂, plan, equation::OZEquation{3, T1, T2, T3}) where {T1, T2, T3}

computes the three dimensional radial fourier transform of F̂ = kf̂(k), returning F = rf(r), where f̂ is the fourier transform of f. it uses the discrete sine tranform provided by the r2r function of FFTW internally.

OrnsteinZernike.solve — Method

OrnsteinZernike.solve — Method

this is a function