Closures

This package defines several closure relations that can be used out of the box. It is straightforward to implement your own closure, see Defining your own closure.

Some closures, for example the Rogers-Young closure, include free parameters that may be fixed by the requirement of thermodynamic consistency. See the Thermodynamic Consistency page for an example.

Some closures use a renormalized indirect correlation function $\gamma^*(r) = \gamma(r) - u_{LR}(r)$ instead of the standard one. Here, $u_{LR}(r)$ is the long range tail of the interaction potential. There are several ways in which the interaction potential can be split into a short-range and long range part, the most common one is the Weeks-Chandler-Andersen construction. In order to use these ...

Implemented Closures

Below is an alphabetical list of implemented closures. We use the notation shown in the Theory section.

OrnsteinZernike.BallonePastoreGalliGazzillo

OrnsteinZernike.BomontBretonnet

OrnsteinZernike.CarbajalTinoko

OrnsteinZernike.CharpentierJackse

OrnsteinZernike.ChoudhuryGhosh

OrnsteinZernike.DuhHaymet

OrnsteinZernike.ExtendedRogersYoung

OrnsteinZernike.HypernettedChain

OrnsteinZernike.Khanpour

OrnsteinZernike.Lee

OrnsteinZernike.MartynovSarkisov

OrnsteinZernike.ModifiedHypernettedChain

OrnsteinZernike.ModifiedVerlet

OrnsteinZernike.PercusYevick

OrnsteinZernike.RogersYoung

OrnsteinZernike.SoftCoreMeanSpherical

OrnsteinZernike.Verlet

OrnsteinZernike.VompeMartynov

OrnsteinZernike.ZerahHansen

OrnsteinZernike.BallonePastoreGalliGazzillo — Type

BallonePastoreGalliGazzillo <: Closure

Implements the Ballone-Pastore-Galli-Gazzillo closure $b(r) = (1 + s \gamma(r))^{1/s} - \gamma(r) - 1 $. Here, $s$ is a free parameter that can be determined with thermodynamic consistency.

Example:

closure = BallonePastoreGalliGazzillo(1.5)

References:

Ballone, P., et al. "Additive and non-additive hard sphere mixtures: Monte Carlo simulation and integral equation results." Molecular Physics 59.2 (1986): 275-290.

OrnsteinZernike.BomontBretonnet — Type

BomontBretonnet <: Closure

Implements the Bomont-Bretonnet closure $b(r) = \sqrt{1+2\gamma^*(r) + f \gamma^*(r)^2} - \gamma^*(r) - 1 $. Here $\gamma^* = \gamma - u_{LR}$ , in which $ u_{LR}$ is the long range tail of the potential, and $f$ is a free parameter that can be determined with thermodynamic consistency.

Example:

closure = BomontBretonnet(0.5)

References:

Bomont, J. M., and J. L. Bretonnet. "A self-consistent integral equation: Bridge function and thermodynamic properties for the Lennard-Jones fluid." The Journal of chemical physics 119.4 (2003): 2188-2191.

OrnsteinZernike.CarbajalTinoko — Type

CarbajalTinoko <: Closure

Implements the Carbajal-Tinoko closure $e(r;\alpha)\left[(2-y(r))e^{y(r)}-2-y(r)\right]/\left(e^{y(r)}-1\right)$ where $e(r;\alpha)=3\exp(\alpha r)$ for $\alpha <0$ and $e(r;\alpha) = 3+\alpha$ otherwise.

Example:

closure = CarbajalTinoko(0.4)

References:

Carbajal-Tinoco, Mauricio D. "Thermodynamically consistent integral equation for soft repulsive spheres." The Journal of chemical physics 128.18 (2008).

OrnsteinZernike.CharpentierJackse — Type

CharpentierJackse <: Closure

Implements the Charpentier-Jackse closure $b(r) = \frac{1}{2\alpha}\left(\sqrt{1+4\alpha\gamma^*(r) } - 2\alpha\gamma^*(r) - 1\right) $. Here $\gamma^* = \gamma - u_{LR}$ , in which $ u_{LR}$ is the long range tail of the potential, and $\alpha$ is a free parameter that can be determined with thermodynamic consistency.

Example:

closure = CharpentierJackse(0.5)

References:

Charpentier, I., and N. Jakse. "Exact numerical derivatives of the pair-correlation function of simple liquids using the tangent linear method." The Journal of Chemical Physics 114.5 (2001): 2284-2292.

OrnsteinZernike.ChoudhuryGhosh — Type

ChoudhuryGhosh <: Closure

Implements the Choudhury-Ghosh closure $b(r) = -\frac{-\gamma^*(r)^2}{2(1+\alpha \gamma^*(r))} $ for $\gamma^*(r) >0$, and $b(r)=-\gamma^*(r)^2/2$ otherwise. Here $\gamma^* = \gamma - u_{LR}$ , in which $ u_{LR}$ is the long range tail of the potential, and $\alpha$ is a free parameter that is determined by an empirical relation.

Example:

α(ρ) = 1.01752 - 0.275ρ # see the reference for this empirical relation
closure = ChoudhuryGhosh(α(0.4))

References:

Choudhury, Niharendu, and Swapan K. Ghosh. "Integral equation theory of Lennard-Jones fluids: A modified Verlet bridge function approach." The Journal of chemical physics 116.19 (2002): 8517-8522.

OrnsteinZernike.DuhHaymet — Type

DuhHaymet <: Closure

Implements the Duh-Haymet closure $b(r) = \frac{-\gamma^*(r)^2}{2\left[1+\left(\frac{5\gamma^*(r)+11}{7\gamma^*(r)+9}\right)\gamma^*(r)\right]} $ for $\gamma^*(r) >0$, and $b(r)=-\gamma^*(r)^2/2$ otherwise. Here $\gamma^* = \gamma - u_{LR}$ , in which $ u_{LR}$ is the long range tail of the potential,

Example:

closure = DuhHaymet()

References:

OrnsteinZernike.ExtendedRogersYoung — Type

ExtendedRogersYoung <: Closure

Implements the extended Rogers-Young closure $b(r) = \ln(a\phi(r)^2 + \phi(r) + 1) - γ(r) $. Here, $\phi(r) = \frac{\exp(f(r)\gamma(r)) - 1}{f(r)}$, and $f(r)=1-\exp(-\alpha r)$, in which $\alpha$ is a free parameter, that may be chosen such that thermodynamic consistency is achieved. Example:

closure = ExtendedRogersYoung(0.5, 0.5) # order is α, a

References: J. Chem. Phys. 128, 184507 (2008)

OrnsteinZernike.HypernettedChain — Type

HypernettedChain

Implements the Hypernetted Chain closure $c(r) = (f(r)+1)\exp(\gamma(r)) - \gamma(r) - 1$, or equivalently $b(r) = 0$.

Example:

closure = HypernettedChain()

OrnsteinZernike.Khanpour — Type

Khanpour <: Closure

Implements the Khanpour closure $b(r) = \frac{1}{\alpha}\ln(1+\alpha\gamma(r)) - \gamma $. Here $\alpha$ is a free parameter that can be determined with thermodynamic consistency.

Example:

closure = Khanpour(0.5)

References:

Khanpour, Mehrdad. "A unified derivation of Percus–Yevick and hyper-netted chain integral equations in liquid state theory." Molecular Physics 120.5 (2022): e2001065.

OrnsteinZernike.Lee — Type

Lee <: Closure

Implements the Lee closure $b(r) = -\frac{\zeta\gamma^*(r)^2}{2} \left( 1- \frac{\phi \alpha \gamma^*(r)}{1 + \alpha\gamma^*(r)} \right) $. Here $\gamma^* = \gamma + ρ f(r)/2$ , in which $ f(r)$ is the Mayer-f function and $\rho$ the density. Additionally, $\zeta$,$\phi$, and, $\alpha$ are free parameters that can be determined with thermodynamic consistency or zero-separation theorems.

Example:

closure = Lee(1.073, 1.816, 1.0, 0.4) # ζ, ϕ, α, ρ

References:

Lee, Lloyd L. "An accurate integral equation theory for hard spheres: Role of the zero‐separation theorems in the closure relation." The Journal of chemical physics 103.21 (1995): 9388-9396.

OrnsteinZernike.MartynovSarkisov — Type

MartynovSarkisov <: Closure

Implements the Martynov-Sarkisov Closure b(r) = -\sqrt{1+2\gamma}-1-\gamma, which is constructed for the 3d hard-sphere liquid.

Example:

closure = MartynovSarkisov()

References:

Martynov, G. A., and G. N. Sarkisov. "Exact equations and the theory of liquids. V." Molecular Physics 49.6 (1983): 1495-1504.

OrnsteinZernike.ModifiedHypernettedChain — Type

ModifiedHypernettedChain <: Closure

Implements the Modified Hypernetted Chain closure $b(r) = b_{HS}(r) $. Here $b_{HS}(r/σ)=\left((a_1+a_2x)(x-a_3)(x-a_4)/(a_3 a_4)\right)^2$ for $x<a_4$ and $b_{HS}(r)=\left(A_1 \exp(-a_5(x-a_4))\sin(A_2(x-a_4))/r\right)^2$ is the hard sphere bridge function found in Malijevský & Labík. The parameters are defined as

\[x = r/σ-1\]

\[A_1 = (a_1+a_2 a_4)(a_4-a_3)(a_4+1)/(A_2 a_3 a_4)\]

\[A_2 = \pi / (a_6 - a_4 - 1)\]

\[a_1 = \eta (1.55707 - 1.85633\eta) / (1-\eta)^2\]

\[a_2 = \eta (1.28127 - 1.82134\eta) / (1-\eta)\]

\[a_3 = (0.74480 - 0.93453\eta)\]

\[a_4 = (1.17102 - 0.68230\eta)\]

\[a_5 = 0.15975/\eta^2\]

\[a_6 = (2.69757 - 0.86987\eta)\]

and $\eta$ is the volume fraction of the hard sphere reference system. This closure only works for single component systems in three dimensions. By default, $\sigma = 1.0$.

Example:

closure = ModifiedHypernettedChain(0.4)
closure = ModifiedHypernettedChain(0.4; sigma=0.8)

References:

Lado, F. "Perturbation correction for the free energy and structure of simple fluids." Physical Review A 8.5 (1973): 2548.

Malijevský, Anatol, and Stanislav Labík. "The bridge function for hard spheres." Molecular Physics 60.3 (1987): 663-669.

OrnsteinZernike.ModifiedVerlet — Type

ModifiedVerlet <: Closure

Implements the modified Verlet Closure $b(r) = -\frac{\gamma^2(r)/2}{1+\alpha \gamma(r)}$. If $\gamma(r)<0$, the closure reads $-\gamma^2/2$. Example:

closure = ModifiedVerlet(0.2)

References:

Verlet, Loup. "Integral equations for classical fluids: I. The hard sphere case." Molecular Physics 41.1 (1980): 183-190.

OrnsteinZernike.PercusYevick — Type

PercusYevick

Implements the Percus-Yevick closure $c(r) = f(r)(1+\gamma(r))$, or equivalently $b(r) = \ln(1 + \gamma(r)) - γ(r)$.

Example:

closure = PercusYevick()

OrnsteinZernike.RogersYoung — Type

RogersYoung <: Closure

Implements the Rogers-Young closure $b(r) = \ln(\frac{\exp(f(r)\gamma(r)) - 1}{f(r)} + 1) - γ(r) $. Here $f(r)=1-\exp(-\alpha r)$, in which $\alpha$ is a free parameter, that may be chosen such that thermodynamic consistency is achieved. Example:

closure = RogersYoung(0.5)

References:

OrnsteinZernike.SoftCoreMeanSpherical — Type

SoftCoreMeanSpherical <: Closure

Implements the soft core mean spherical closure $b(r) = \ln(\gamma^*(r) + 1) - \gamma^*(r)$. Here $\gamma^* = \gamma - u_{LR}$ , in which $u_{LR}$ is the long range tail of the potential.

Example:

closure = SoftCoreMeanSpherical()

References:

OrnsteinZernike.Verlet — Type

Verlet <: Closure

Implements the Verlet Closure $b(r) = -\frac{A\gamma^2(r)/2}{1+B \gamma(r)/2}$, where by default $A=1$ and $B=8/5$. These values are tuned by the virial coefficients of the 3d hard sphere liquid.

Example:

closure = Verlet()
closure = Verlet(A=3.0, B=4.0)

References:

Verlet, Loup. "Integral equations for classical fluids: I. The hard sphere case." Molecular Physics 41.1 (1980): 183-190.

OrnsteinZernike.VompeMartynov — Type

VompeMartynov <: Closure

Implements the Vompe-Martynov closure $b(r) = \sqrt{1+2\gamma^*(r) } - \gamma^*(r) - 1 $. Here $\gamma^* = \gamma - u_{LR}$ , in which $ u_{LR}$ is the long range tail of the potential, and $\alpha$ is a free parameter that can be determined with thermodynamic consistency.

Example:

closure = VompeMartynov()

References:

Vompe, A. G., and G. A. Martynov. "The bridge function expansion and the self‐consistency problem of the Ornstein–Zernike equation solution." The Journal of chemical physics 100.7 (1994): 5249-5258.

OrnsteinZernike.ZerahHansen — Type

ZerahHansen <: Closure

Implements the Zerah-Hansen (HMSA) (HNC-SMSA) closure $b(r) = \ln(\frac{\exp(f(r)\gamma^*(r)) - 1}{f(r)} + 1) - γ^*(r) $. Here $\gamma^* = \gamma - u_{LR}$ , in which $ u_{LR}$ is the long range tail of the potential, and $f(r)=1-\exp(-\alpha r)$, in which $\alpha$ is a free parameter, that may be chosen such that thermodynamic consistency is achieved.

Example:

closure = ZerahHansen(0.5)

References:

Zerah, Gilles, and Jean‐Pierre Hansen. "Self‐consistent integral equations for fluid pair distribution functions: Another attempt." The Journal of chemical physics 84.4 (1986): 2336-2343.